requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.2.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-loading-spinners==1.0.3
dash-table==5.0.0
dash_ag_grid==31.2.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==23.0.0
h5py==3.12.1
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.3
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
quantstats-lumi==0.3.3
redis==5.2.0
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rq==1.16.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
ta==0.11.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.16.0
yfinance==0.2.48
zipp==3.20.2
